+++ tools/explorerEval/tool.js
@@ -145,7 +145,7 @@
         }
 
         const q=1000/total;        
-        const [w,d,l]=[explorerItem.white*q,(explorerItem.draws+1)*q,explorerItem.black*q];
+        const [w,d,l]=[explorerItem.white*q,Math.max(explorerItem.draws,1)*q,explorerItem.black*q];
         const sharpness = Math.round(Math.min(w,l)/50*333/d*1/(1+Math.exp(-(w+l)/1000)));
         if (sharpness&&Number.isFinite(sharpness)) {
           const sharpnessTitle = trans.pluralSame('sharpnessTitle',sharpness);